Automating virtual machines backup
The main file is vmbackup.sh, a shell script which makes the real thing. It uses vim-cmd
commands to shutdown the appropriate VM, clone locally its files, restart the VM again and
copying the local copy to a NAS in the background.

The other files are need to allow crond to work properly and the changes to system files
to be restored after a possible host restart, which usually restore most files to system
defaults.